+++ test/performance/ranges/views/view_translate_1D_benchmark.cpp
@@ -11,7 +11,7 @@
 #include <benchmark/benchmark.h>
 
 #include <bio/alphabet/nucleotide/dna4.hpp>
-#include <bio/ranges/views/to.hpp>
+#include <bio/ranges/to.hpp>
 #include <bio/ranges/views/translate.hpp>
 #include <bio/test/performance/sequence_generator.hpp>
 #include <bio/test/seqan2.hpp>
@@ -46,7 +46,7 @@ void sequential_read(benchmark::State & state)
     if constexpr (std::is_same_v<tag_t, baseline_tag>)
     {
         bio::alphabet::aa27_vector translated_aa_sequence = dna_sequence | bio::ranges::views::translate_single
-                                                                  | bio::ranges::views::to<bio::alphabet::aa27_vector>();
+                                                                  | bio::ranges::to<bio::alphabet::aa27_vector>();
         sequential_read_impl(state, translated_aa_sequence);
     }
     else if constexpr (std::is_same_v<tag_t, translate_tag>)
@@ -87,7 +87,7 @@ void random_access(benchmark::State & state)
     if constexpr (std::is_same_v<tag_t, baseline_tag>)
     {
         bio::alphabet::aa27_vector translated_aa_sequence = dna_sequence | bio::ranges::views::translate_single
-                                                                  | bio::ranges::views::to<bio::alphabet::aa27_vector>();
+                                                                  | bio::ranges::to<bio::alphabet::aa27_vector>();
         random_access_impl(state, translated_aa_sequence, access_positions);
     }
     else
@@ -111,7 +111,7 @@ void copy_impl(benchmark::State & state, std::vector<bio::alphabet::dna4> const 
     {
         bio::alphabet::aa27_vector translated_aa_sequence{};
         benchmark::DoNotOptimize(translated_aa_sequence = dna_sequence | adaptor
-                                                                       | bio::ranges::views::to<bio::alphabet::aa27_vector>());
+                                                                       | bio::ranges::to<bio::alphabet::aa27_vector>());
     }
 }