+++ test/snippet/ranges/views/range_view_all_retransform.cpp
@@ -1,6 +1,6 @@
 #include <bio/alphabet/nucleotide/dna4.hpp>
 #include <bio/ranges/views/complement.hpp>
-#include <bio/ranges/views/to.hpp>
+#include <bio/ranges/to.hpp>
 #include <ranges>
 
 int main()
@@ -11,10 +11,10 @@ int main()
     auto vec_view2 = bio::views::complement(vec);
 
     // re-convert to container
-    bio::alphabet::dna4_vector complemented = vec_view2 | bio::views::to<bio::alphabet::dna4_vector>();
+    bio::alphabet::dna4_vector complemented = vec_view2 | bio::ranges::to<bio::alphabet::dna4_vector>();
     assert(complemented == "TGCCAG"_dna4);
 
     // also possible in one step
-    bio::alphabet::dna4_vector reversed = vec | std::views::reverse | bio::views::to<bio::alphabet::dna4_vector>();
+    bio::alphabet::dna4_vector reversed = vec | std::views::reverse | bio::ranges::to<bio::alphabet::dna4_vector>();
     assert(reversed == "CTGGCA"_dna4);
 }