+++ test/unit/ranges/views/view_char_to_test.cpp
@@ -13,8 +13,8 @@
 #include <algorithm>
 #include <bio/alphabet/nucleotide/dna5.hpp>
 #include <bio/ranges/concept.hpp>
+#include <bio/ranges/to.hpp>
 #include <bio/ranges/views/char_to.hpp>
-#include <bio/ranges/views/to.hpp>
 #include <bio/test/expect_range_eq.hpp>
 #include <ranges>
 
@@ -27,18 +27,18 @@ TEST(view_char_to, basic)
 
     // pipe notation
     bio::alphabet::dna5_vector v =
-      vec | bio::ranges::views::char_to<bio::alphabet::dna5> | bio::ranges::views::to<std::vector>();
+      vec | bio::ranges::views::char_to<bio::alphabet::dna5> | bio::ranges::to<std::vector>();
     EXPECT_EQ(cmp, v);
 
     // function notation
     bio::alphabet::dna5_vector v2(bio::ranges::views::char_to<bio::alphabet::dna5>(vec) |
-                                  bio::ranges::views::to<std::vector>());
+                                  bio::ranges::to<std::vector>());
     EXPECT_EQ(cmp, v2);
 
     // combinability
     bio::alphabet::dna5_vector cmp2{"ATAGTTTCA"_dna5};
-    bio::alphabet::dna5_vector v3 = vec | bio::ranges::views::char_to<bio::alphabet::dna5> | std::views::reverse |
-                                    bio::ranges::views::to<std::vector>();
+    bio::alphabet::dna5_vector v3 =
+      vec | bio::ranges::views::char_to<bio::alphabet::dna5> | std::views::reverse | bio::ranges::to<std::vector>();
     EXPECT_EQ(cmp2, v3);
 }
 
@@ -47,7 +47,7 @@ TEST(view_char_to, deep_view)
     std::vector<std::string> foo{"ACGTA", "TGCAT"};
 
     std::vector<bio::alphabet::dna5_vector> v = foo | bio::ranges::views::char_to<bio::alphabet::dna5> |
-                                                bio::ranges::views::to<std::vector<bio::alphabet::dna5_vector>>();
+                                                bio::ranges::to<std::vector<bio::alphabet::dna5_vector>>();
 
     ASSERT_EQ(size(v), 2u);
     EXPECT_RANGE_EQ(v[0], "ACGTA"_dna5);