+++ test/unit/ranges/views/view_move_test.cpp
@@ -13,9 +13,9 @@
 #include <algorithm>
 #include <bio/alphabet/nucleotide/dna5.hpp>
 #include <bio/ranges/concept.hpp>
+#include <bio/ranges/to.hpp>
 #include <bio/ranges/views/complement.hpp>
 #include <bio/ranges/views/move.hpp>
-#include <bio/ranges/views/to.hpp>
 #include <bio/ranges/views/to_lower.hpp>
 #include <bio/test/expect_range_eq.hpp>
 #include <ranges>
@@ -38,7 +38,7 @@ TEST(view_move, basic)
     bio::alphabet::dna5_vector vec2{"ACGTA"_dna5};
     bio::alphabet::dna5_vector v3 = vec2 | bio::ranges::views::complement |
                                     bio::ranges::views::move // NOP, because already temporaries
-                                    | bio::ranges::views::to<std::vector>();
+                                    | bio::ranges::to<std::vector>();
     EXPECT_EQ("TGCAT"_dna5, v3);
 }