+++ test/unit/ranges/views/view_slice_test.cpp
@@ -17,9 +17,9 @@
 #include <algorithm>
 #include <bio/ranges/concept.hpp>
 #include <bio/ranges/container/concept.hpp>
+#include <bio/ranges/to.hpp>
 #include <bio/ranges/views/single_pass_input.hpp>
 #include <bio/ranges/views/slice.hpp>
-#include <bio/ranges/views/to.hpp>
 #include <bio/test/expect_range_eq.hpp>
 #include <bio/test/expect_same_type.hpp>
 #include <concepts>