+++ CHANGELOG.md
@@ -9,12 +9,18 @@ bug-fixes (not all bug-fixes will be listed).
 
 ## Features
 
+* Added bio::meta::vtag, bio::meta::ttag and bio::meta::priority_tag.
+
 ## Bug-fixes
 
 ## API
 
 * This release is not yet API-stable.
 * This release is *mostly* backwards compatible with 0.5.0.
+
+### Misc changes
+
+* `bio::ranges::views::to()` has been renamed to `bio::ranges::to()` to be in line with C++23.
 
 # 0.5.0